refactor(dashboard): simplify pass on team-operator-wiring branch
- Extract shared formatRelativeAgo into src/lib/format.ts; drop the two
  duplicate formatAgo copies in operator/page.tsx and CustomerRow.tsx.
- Extract readApiError into src/lib/errors.ts; drop duplicate in
  ClearDemoButton.tsx and team/members/page.tsx.
- formatSyncAgo now delegates to formatRelativeAgo with the
  'never synced' fallback, keeping the public API + test contract.
- Trim meta-narration comments from lib/team.ts.

All 78 vitest tests green; tsc --noEmit clean.
